reconstructing the file directory
Move arknight_operator_dataset into the datasets folder just in case you need to import more than one dataset in the future, so that the project structure is more organized.

Replace the static dataset path with a dynamic one to minimize errors when running on other computers. Python v3 already considers the runner directory as the root, so there is no need to use a static path.